rename explorer to library and cleanup grid
