--- assets/src/edit-story/components/explorer/index.js
@@ -1,3 +1,0 @@
-export { default } from './explorer';
-export { default as ExplorerTabs } from './explorerTabs';
-export { default as ExplorerProvider } from './explorerProvider';
