+++ assets/src/edit-story/components/library/index.js
@@ -1,0 +1,3 @@
+export { default } from './library';
+export { default as LibraryTabs } from './libraryTabs';
+export { default as LibraryProvider } from './libraryProvider';
